Move server objects to Archon.AspNetCore
