+++ aspnetcore/aspnetcore.csproj
@@ -1,15 +1,29 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
+    <Description>Middleware and helpers to make working with ASP.NET MVC easier.</Description>
+    <Authors>Archon Information Systems, LLC</Authors>
     <TargetFramework>netstandard1.3</TargetFramework>
     <AssemblyName>Archon.AspNetCore</AssemblyName>
     <RootNamespace>Archon.AspNetCore</RootNamespace>
+    <PackageTags>utils;api;http;model binders;middleware</PackageTags>
+    <PackageProjectUrl>https://github.com/civicsource/http</PackageProjectUrl>
+    <PackageLicenseUrl>https://github.com/civicsource/http/blob/master/LICENSE</PackageLicenseUrl>
+    <Version>1.0.0</Version>
+    <RepositoryUrl>https://github.com/civicsource/http.git</RepositoryUrl>
+    <RepositoryType>git</RepositoryType>
+    <Product>Archon Http Utilities</Product>
   </PropertyGroup>
 
+  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
+    <TreatWarningsAsErrors>True</TreatWarningsAsErrors>
+    <TreatSpecificWarningsAsErrors />
+  </PropertyGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.AspNetCore.Diagnostics" Version="1.1.2" />
     <PackageReference Include="Microsoft.AspNetCore.Http.Abstractions" Version="1.1.2" />
     <PackageReference Include="Microsoft.AspNetCore.Mvc.Abstractions" Version="1.1.3" />
+    <PackageReference Include="Newtonsoft.Json" Version="10.0.3" />
   </ItemGroup>
 
 </Project>
